Ordering_tests/GPT_Gemini/requirements.txt
pandas
google-generativeai
qatch
requests
dotenv
sqlparse